+++ scr/controllers/partner.js
@@ -1,6 +1,8 @@
 const createError = require('http-errors');
 const { isValidObjectId } = require('mongoose');
+const Customer = require('../models/customer');
 const Partner = require('../models/partner');
+const Admin = require('../models/admin');
 const sendMail = require('../utils/email/sendEmail');
 const utils = require('../utils/index');
 const {
@@ -49,18 +51,25 @@ const getAllPartners = async (req, res, next) => {
 
 const getPartner = async (req, res, next) => {
   const id = req.params.id;
+  const userId = req.userId;
 
   if (isValidObjectId(id)) {
-    try {
-      const partner = await Partner.findById(id).catch((error) => next(error));
+    if (id === userId) {
+      try {
+        const partner = await Partner.findById(id).catch((error) =>
+          next(error),
+        );
 
-      if (partner == null) {
-        next(createError.NotFound("This Partner doesn't exist"));
-      } else if (partner) {
-        res.json({ data: partner, success: true, error: null });
+        if (partner == null) {
+          next(createError.NotFound("This Partner doesn't exist"));
+        } else if (partner) {
+          res.json({ data: partner, success: true, error: null });
+        }
+      } catch (error) {
+        next(error);
       }
-    } catch (error) {
-      next(error);
+    } else {
+      next(createError.Unauthorized('Unauthorized'));
     }
   } else {
     next(createError.BadRequest('Invalid ID'));
@@ -70,26 +79,31 @@ const getPartner = async (req, res, next) => {
 const updatePartner = async (req, res, next) => {
   const id = req.params.id;
   const updated = req.body;
+  const userId = req.userId;
 
   if (isValidObjectId(id)) {
     if (!utils.isObjctEmpty(updated)) {
-      try {
-        const updatedPartner = await Partner.findByIdAndUpdate(id, updated, {
-          new: true,
-        }).catch((error) => next(error));
+      if (id === userId) {
+        try {
+          const updatedPartner = await Partner.findByIdAndUpdate(id, updated, {
+            new: true,
+          }).catch((error) => next(error));
 
-        if (updatedPartner) {
-          res.json({
-            message: 'Updated successful',
-            data: updatedPartner,
-            success: true,
-            error: null,
-          });
-        } else {
-          throw createError.NotFound("This partner doesn't exist");
+          if (updatedPartner) {
+            res.json({
+              message: 'Updated successful',
+              data: updatedPartner,
+              success: true,
+              error: null,
+            });
+          } else {
+            throw createError.NotFound("This partner doesn't exist");
+          }
+        } catch (error) {
+          next(error);
         }
-      } catch (error) {
-        next(error);
+      } else {
+        next(createError.Unauthorized('Unauthorized'));
       }
     } else {
       next(createError.NotAcceptable('Please , mention the fields to update'));
@@ -103,10 +117,16 @@ const createPartner = async (req, res, next) => {
   try {
     const result = await partnerRegisterSchema.validateAsync(req.body);
 
-    const isExist = await Partner.findOne({ userName: result.userName }).catch(
-      (error) => next(error),
-    );
-
+    const isExist =
+      (await Customer.findOne({ userName: result.userName }).catch((error) =>
+        next(error),
+      )) ||
+      (await Partner.findOne({ userName: result.userName }).catch((error) =>
+        next(error),
+      )) ||
+      (await Admin.findOne({ userName: result.userName }).catch((error) =>
+        next(error),
+      ));
     if (isExist) {
       throw createError.Conflict('This partner is already exist');
     }
@@ -140,24 +160,30 @@ const createPartner = async (req, res, next) => {
 
 const deletePartner = async (req, res, next) => {
   const id = req.params.id;
-  if (isValidObjectId(id)) {
-    try {
-      const deletedPartner = await Partner.findByIdAndDelete(id).catch(
-        (error) => next(error),
-      );
+  const userId = req.userId;
 
-      if (deletedPartner) {
-        res.json({
-          message: 'Deleted successful',
-          data: deletedPartner,
-          success: true,
-          error: null,
-        });
-      } else {
-        throw createError.NotFound("This partner doesn't exist");
+  if (isValidObjectId(id)) {
+    if (id === userId) {
+      try {
+        const deletedPartner = await Partner.findByIdAndDelete(id).catch(
+          (error) => next(error),
+        );
+
+        if (deletedPartner) {
+          res.json({
+            message: 'Deleted successful',
+            data: deletedPartner,
+            success: true,
+            error: null,
+          });
+        } else {
+          throw createError.NotFound("This partner doesn't exist");
+        }
+      } catch (error) {
+        next(error);
       }
-    } catch (error) {
-      next(error);
+    } else {
+      next(createError.Unauthorized('Unauthorized'));
     }
   } else {
     next(createError.BadRequest('Invalid ID'));
